Fixed scripts that open FLEx.
Changed open_new_project to use restart_flex() on green screen error,
instead of Ryan's script.
Changed restart_flex to just quit if the project isn't found, instead of
trying open_new_project, because otherwise it would just loop forever.
Put the "exit()" back in 1_open_flex_existing, because otherwise we don't
know when to write a failure to the log. (The error message in the terminal
when we exit that way is benign.)