Factored out some default constants, added Wait to Regionplus
Default time, success message, and options for giving up and
restarting arguments to wrapper functions are now set in
the constructor, in both Regionplus and TestHelper. Regionplus
copies defaults from its TestHelper when the TestHelper is set.

The Wait wrapper previously added to TestHelper has now been
added to Regionplus as well.