Add image_grabber
This script generates new scripts to run to facilitate recapture images in all tests.
This can be used to port scripts from one OS to another OS.
This can be used when a systematic change alters a majority of images in tests on a single OS.
Run the script itself via 'sikuli-ide -r image_grabber.sikuli" to generate the files.
Modify existing runall scripts to point at these files.