Move and update image_updater
Moved image_updater to the ITF folder so the script wouldn't recursively
make subfolders in the image_updater folder forever.

Renamed sikuli script and class to show the intent to update and not start
fresh. Renamed interactive capture method to be more explanatory. Removed
some testing artifacts to make the code more precise.