Extend logging to include successes and console output
Instead of logging only errors, successful images would help be a sanity
check for testers to see all steps in the tests. This involves adding a
type of log entry (Success and Fail), getting a preemptive screenshot, and
updating the logs to reflect these attributes.
Printing the test_helper activity enables the testers to use the console on
the CI server or interactive sessions to see all of the data without
opening the error_log file.