Update localization tests images and rename localization folders
Images and regions needed to be updated for FLEX 8.2.
Only a couple images needed to be changed. A switch was added in the
generic test to capture images for each language to compare manually.
An enlish localization was added as a sanity check and as a reset
capability in the event that a non-English localization doesn't finish.
The folders were changed in preparation for the xmlrunner capability
so tests would not have the names "test_test_*localization".