Automate steps for screen related tasks and add fix miscellaneous problems
Changing the screen resolution, blackout, and lock steps plus changing the
adding the french keyboad step were done manually after the install auto
was performed.
These steps were incorporated in 95a480d, but were not effective in the
provisioner because it is run as the root user.
The necessary commands are input redirected to be executed as the vagrant
user because the user is automatically logged in and has the X session
available we will use for testing.
As far as miscellaneous problems, flex wouldn't install unless apt-get
update was ran. Also, the VM ran too slow and missed some images.
Use the vagrant option "--provision-with" to specify use the provisioners.