Convert test output to feed Jenkins
Tests were run with shell scripts, but successes and failures were not
in a format Jenkins could use.
A test class was written to include all of the tests and old files
were removed. Other small miscellaneous changes were made.
These tests should be easier to maintain and output directly to xml.