Make test_helper transparent
Test_helper was originally written for use with the sikuli IDE only.
Test_helper has been changed to be a module that extends the native
capabilities of sikuli with less maintenance burden.
Major revisions include:
    - Converting test_helper to a module instead of a class
        (This removes the custom logging feature in favor of unit testing)
    - Adding Debug Workaround Methods
    - Including display_log.py as a function and adding trigger methods
    - Removing a_setup and associated settings
    - Converting scripts to the new format
This was done to keep users from needing to initialize a test helper and
modifying the IDE code.