Reimplement image_updater to be a function of test_helper
Image_updater was a long drawn out process that involved recreating the
python files into a temporary location and using hotkeys to manuever
between actions.
The new image_updater takes advantage of the screen_logger's find failed
awareness and uses this as the place to replace the problematic images --
one by one.
This will make the updating process even simpler for people to use.