Correct image_updating bugs
Test_helper was not able to be used across multiple instances of files in
the IDE because the import mechanism of python only imports once.
The base functionality of test_helper was added to a function and the
import mechanism was decorated to rerun this init function every time
test_helper is imported. The ability to add the common images folder was
also added.
These capabilities will aid in script writing and maitenance.