Add dragDrop logging and better common image path process
The dragDrop function uses two images to operate, but the screen_logger
only used one image. If the common image path didn't exist an error is
shown even with MAKE as the default behavior because of there the print
statement is.
DragDrop checks are added to perform various data structuring and extra
logic is added in the FindFailed block. The common image behavior error
was made a warning and the order prioritized.
The both make test_helper awesome.