Fix minor problems
Non-image (i.e. wait(1) would break simple compare and not having a
favorites region was also problematic. The vagrant file would choke when
adding the ~/.Sikulix/Lib folder. XMLrunner was old.
A text_log flag is set when the first parameter is a number of float to
short circuit the similarity check. The favorite regions was checked. The
Vagrantfile was chenged to add the folders recursively if they don't exits.
XMLrunner was updated to be used with Python 2.7.
The should help handle more use normal cases.